Move the Tei language parsing into a standalone object
